+++ form/src/main/java/feign/form/multipart/PartResolver.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright © 2012 The Feign Authors (feign@commonhaus.dev)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package feign.form.multipart;
+
+import feign.Request;
+import feign.codec.EncodeException;
+import java.util.stream.Stream;
+
+/**
+ * A strategy for resolving {@link PartMetadata} into one or more {@link Request.Body} instances.
+ */
+@FunctionalInterface
+public interface PartResolver {
+  /**
+   * Resolves the given {@link PartMetadata} into one or more {@link Request.Body} instances.
+   *
+   * @param partMetadata the metadata of the part to resolve
+   * @param chain the chain of resolvers to delegate to if this resolver cannot handle the part
+   * @return a stream of {@link Request.Body} instances representing the resolved part
+   * @throws EncodeException if an error occurs during encoding the part
+   */
+  Stream<Request.Body> resolve(PartMetadata partMetadata, PartResolverChain chain)
+      throws EncodeException;
+
+  /**
+   * Determines whether this resolver supports the given {@link PartMetadata}. By default, it
+   * returns {@code true}.
+   *
+   * @param partMetadata the metadata of the part to check for support
+   * @return {@code true} if this resolver supports the given part metadata, {@code false} otherwise
+   */
+  default boolean supports(PartMetadata partMetadata) {
+    return true;
+  }
+}